refactor!: remove top level reexport and rename to methods / types
