+++ examples/api_trait_implementation.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use frankenstein::api_params::SendMessageParams;
+use frankenstein::methods::SendMessageParams;
 use frankenstein::response::ErrorResponse;
 use frankenstein::TelegramApi;
 use isahc::prelude::*;
