+++ src/trait_sync.rs
@@ -4,14 +4,14 @@ use crate::games::GameHighScore;
 use crate::inline_mode::{PreparedInlineMessage, SentWebAppMessage};
 use crate::input_file::HasInputFile;
 use crate::input_media::{InputMedia, MediaGroupInputMedia};
-use crate::objects::{
+use crate::payments::StarTransactions;
+use crate::response::{MessageOrBool, MethodResponse};
+use crate::stickers::{Gifts, Sticker, StickerSet};
+use crate::types::{
     BotCommand, BotDescription, BotName, BotShortDescription, BusinessConnection,
     ChatAdministratorRights, ChatFullInfo, ChatInviteLink, ChatMember, File, ForumTopic,
     MenuButton, Message, MessageId, Poll, User, UserChatBoosts, UserProfilePhotos,
 };
-use crate::payments::StarTransactions;
-use crate::response::{MessageOrBool, MethodResponse};
-use crate::stickers::{Gifts, Sticker, StickerSet};
 use crate::updates::{Update, WebhookInfo};
 
 macro_rules! request {
@@ -21,7 +21,7 @@ macro_rules! request {
             #[inline(always)]
             fn [<$name:snake>] (
                 &self,
-                params: &crate::api_params::[<$name:camel Params>],
+                params: &crate::methods::[<$name:camel Params>],
             ) -> Result<MethodResponse<$return>, Self::Error> {
                 self.request(stringify!($name), Some(params))
             }
@@ -52,7 +52,7 @@ macro_rules! request_f {
             #[doc = "Call the `" $name "` method.\n\nSee <https://core.telegram.org/bots/api#" $name:lower ">."]
             fn [<$name:snake>] (
                 &self,
-                params: &crate::api_params::[<$name:camel Params>],
+                params: &crate::methods::[<$name:camel Params>],
             ) -> Result<MethodResponse<$return>, Self::Error> {
                 let mut files = Vec::new();
                 $(
@@ -86,7 +86,7 @@ pub trait TelegramApi {
 
     fn send_media_group(
         &self,
-        params: &crate::api_params::SendMediaGroupParams,
+        params: &crate::methods::SendMediaGroupParams,
     ) -> Result<MethodResponse<Vec<Message>>, Self::Error> {
         let mut files = Vec::new();
 
@@ -164,7 +164,7 @@ pub trait TelegramApi {
 
     fn set_chat_photo(
         &self,
-        params: &crate::api_params::SetChatPhotoParams,
+        params: &crate::methods::SetChatPhotoParams,
     ) -> Result<MethodResponse<bool>, Self::Error> {
         let photo = &params.photo;
         self.request_with_form_data("setChatPhoto", params, vec![("photo", photo.path.clone())])
@@ -213,7 +213,7 @@ pub trait TelegramApi {
 
     fn edit_message_media(
         &self,
-        params: &crate::api_params::EditMessageMediaParams,
+        params: &crate::methods::EditMessageMediaParams,
     ) -> Result<MethodResponse<MessageOrBool>, Self::Error> {
         let mut files = Vec::new();
 
@@ -262,7 +262,7 @@ pub trait TelegramApi {
 
     fn upload_sticker_file(
         &self,
-        params: &crate::api_params::UploadStickerFileParams,
+        params: &crate::methods::UploadStickerFileParams,
     ) -> Result<MethodResponse<File>, Self::Error> {
         let sticker = &params.sticker;
         self.request_with_form_data(
@@ -274,7 +274,7 @@ pub trait TelegramApi {
 
     fn create_new_sticker_set(
         &self,
-        params: &crate::api_params::CreateNewStickerSetParams,
+        params: &crate::methods::CreateNewStickerSetParams,
     ) -> Result<MethodResponse<bool>, Self::Error> {
         let mut files = Vec::new();
 
@@ -297,7 +297,7 @@ pub trait TelegramApi {
 
     fn add_sticker_to_set(
         &self,
-        params: &crate::api_params::AddStickerToSetParams,
+        params: &crate::methods::AddStickerToSetParams,
     ) -> Result<MethodResponse<bool>, Self::Error> {
         let mut files = Vec::new();
         if let Some(file) = params.sticker.sticker.clone_path() {
